sync container offset and input scroll
We are rendering the fancy highlighted tokens on top of the input. This
still allow us to just type in the input but get syntax highlighting on
top.

However, when the input contents is getting big, it will result in a
scrollable input. The moment you start scrolling (or navigate within the
input using arrow keys), then the layer on top will not be positioned
correctly.

To solve this, we sync the scroll position with an offset and use a
rAF-loop to keep things in sync.

As an optimization, the rAF-loop only happens if the input has focus.