make `ERROR` an explicit `EvaluationResult` kind
This makes sure that an `ERROR` is an actual value. Before this, we were
throwing errors, which means that if you reference a cell that resulted
in an error, your cell will also be an error.

There are some functions (e.g.: `COUNT()`) that can ignore empty cells,
and cells with errors. By making `ERROR` a valid `EvaluationResult`, we
can still compute the `COUNT()` instead of also erroring out.

In addition, this commit gets rid of the `ComputationResult` concept,
and uses teh `EvaluationResult` directly.

Now all errors all values.

Last but not least, some error messages are updated / improved, because
we can handle them more correctly now.