make `EMPTY` a first class type
When a cell is referenced, and the cell is empty it gets the `EMPTY`
type.

Initial thought was to make it a `NUMBER` with value `0`. Which works in
places where `0` is safe to use (E.g.: `SUM(…)`). But when using it in
places where it's not (e.g.: `PRODUCT(…)`, here we would expect `1`) we
want a different value.

Leaving out the value entirely would also not be correct. E.g.:
`PI(B2)`, if `B2` is empty, and it didn't resolve in any value, then
the function `PI(…)` wouldn't see any arguments, but there _is_ an
argument.