Add function signatures (#1)
This PR adds function signatures to each function. This includes each
parameter, whether it's optional or not, whether it's variadic and what
the potential types are.

It also adds some information via jsdoc-like `@` tags. This includes the
`@description The description of the function goes here` and `@param
name The description of the parameter goes here` tags.

This PR allows us to:

1. Remove the (duplicate) validation we used to have
2. Make the errors more consistent
3. Provide better documentation for the functions when looking at
suggestions

Exposing new functions now look like this:
```ts
export const IMUL = expose(
  `
    @description The result of 32-bit multiplication of two numbers.
    @param x First number
    @param y Second number
    IMUL(x: NUMBER, y: NUMBER)
  `,
  (x: EvaluationResultNumber, y: EvaluationResultNumber) => {
    return { kind: EvaluationResultKind.NUMBER, value: Math.imul(x.value, y.value) }
  },
)
```