add special built in `INHERIT_FORMULA` function
This function has super powers (it needs access to some more private
information), and allows you to inherit a formula of a different cell.

All cell references and range references will be updated relative to
your current cell you are using the formula in. This means that there
is a single source of truth for a formula.

Making changes to the single cell will reflect in the other cells
automatically.

Some other features:
- As expected, locked cells, e.g. `A$1` will keep the location locked,
  and only the none-locked parts will update the location relative to
  the current cell.
- In the UI, inherited dependencies will be highlighted as well as the
  main dependency.